Move pulse utils, which are Schedule -> Schedule functions, to a reschedule file
